[Swift Build] Isolate embedded resource task planning
Move embedded-resource discovery, validation, and per-architecture object task construction into dedicated SourcesTaskProducer support. This keeps the main source producer focused on adding the generated accessor and linker inputs.\n\nRestrict the object-resource metadata and custom spec entry points to package scope so the implementation does not require SWBCore to publicly re-export SWBUtil.
