+++ Sources/SWBTaskConstruction/CMakeLists.txt
@@ -21,6 +21,7 @@ add_library(SWBTaskConstruction
   TaskProducers/BuildPhaseTaskProducers/AppleScriptTaskProducer.swift
   TaskProducers/BuildPhaseTaskProducers/BuildRuleTaskProducer.swift
   TaskProducers/BuildPhaseTaskProducers/CopyFilesTaskProducer.swift
+  TaskProducers/BuildPhaseTaskProducers/EmbeddedResourceTaskSupport.swift
   TaskProducers/BuildPhaseTaskProducers/FilesBasedBuildPhaseTaskProducer.swift
   TaskProducers/BuildPhaseTaskProducers/HeadersTaskProducer.swift
   TaskProducers/BuildPhaseTaskProducers/ResourcesTaskProducer.swift
